backend/pyproject.toml
[project]
name = "lucknow-tech-events-backend"
version = "0.1.0"
description = "Backend for Lucknow tech events aggregator"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "sqlalchemy>=2.0",
  "asyncpg>=0.29",
  "psycopg2-binary>=2.9",
  "alembic>=1.13",
  "httpx>=0.27",
  "playwright>=1.47",
  "celery>=5.4",
  "redis>=5.0",
  "structlog>=24.4",
  "python-jose[cryptography]>=3.3",
  "passlib[bcrypt]>=1.7",
  "slowapi>=0.1.9",
  "icalendar>=5.0",
  "python-slugify>=8.0",
  "python-dateutil>=2.9",
  "pytz>=2024.1",
  "google-genai>=1.0.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]

